Extract dashboard.db_queries to api.statistics.db_queries
