+++ api/statistics/db_queries/node_availability.py
